chore: reorganize project structure by moving files to appropriate packages
