+++ src/main/kotlin/ecommerce/repository/CollectionProductStore.kt
@@ -1,4 +1,4 @@
-package ecommerce
+package ecommerce.repository
 
 import ecommerce.model.Product
 import org.springframework.stereotype.Repository
