+++ src/main/kotlin/ecommerce/repository/ProductStore.kt
@@ -1,4 +1,4 @@
-package ecommerce
+package ecommerce.repository
 
 import ecommerce.model.Product
